requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
certifi==2023.7.22
charset-normalizer==3.3.2
Deprecated==1.2.14
idna==3.4
importlib-resources==6.1.0
limits==3.6.0
multidict==6.0.4
packaging==23.2
pamqp==3.2.1
pylogbeat==2.0.1
python-dotenv==1.0.0
python-logstash-async==2.7.2
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
yarl==1.9.2
